shift chartDataFunctionFromPoints into TableCatalogItem
