+++ lib/Models/TableCatalogItem.js
@@ -13,6 +13,7 @@ var when = require('terriajs-cesium/Source/ThirdParty/when');
 var TimeIntervalCollection = require('terriajs-cesium/Source/Core/TimeIntervalCollection');
 
 var CatalogItem = require('./CatalogItem');
+var ChartData = require('../Charts/ChartData');
 var inherit = require('../Core/inherit');
 var overrideProperty = require('../Core/overrideProperty');
 var Polling = require('./Polling');
@@ -856,6 +857,32 @@ TableCatalogItem.prototype.getNextColor = function() {
         }
     }
     return colors[0];
+};
+
+/**
+ * Returns a {@link ChartData} object for the TableCatalogItem.  See ChartPanel.jsx for an example.
+ *
+ * @returns {ChartData}
+ */
+TableCatalogItem.prototype.chartData = function() {
+    const item = this;
+    const xColumn = item.xAxis;
+    const yColumns = item.tableStructure.columnsByType[VarType.SCALAR].filter(column => column.isActive);
+    if (yColumns.length > 0) {
+        const yColumnNumbers = yColumns.map(yColumn => item.tableStructure.columns.indexOf(yColumn));
+        const pointArrays = item.tableStructure.toPointArrays(xColumn, yColumns);
+        return pointArrays.map((points, index) =>
+            new ChartData(points, {
+                id: item.uniqueId + '-' + yColumnNumbers[index],
+                name: yColumns[index].name,
+                categoryName: item.name,
+                units: yColumns[index].units,
+                color: yColumns[index].color,
+                yAxisMin: yColumns[index].yAxisMin,
+                yAxisMax: yColumns[index].yAxisMax
+            })
+        );
+    }
 };
 
 /**
